Added account verfication to manager layer, all tests now pass.

+++ CECS 491B/TrialByFire.Tresearch/TrialByFire.Tresearch.Managers/Implementations/AccountDeletionManager.cs
@@ -11,6 +11,7 @@ using TrialByFire.Tresearch.Exceptions;
 using TrialByFire.Tresearch.Managers.Contracts;
 using TrialByFire.Tresearch.Models;
 using TrialByFire.Tresearch.Models.Contracts;
+using TrialByFire.Tresearch.Models.Implementations;
 using TrialByFire.Tresearch.Services.Contracts;
 
 namespace TrialByFire.Tresearch.Managers.Implementations
@@ -22,15 +23,16 @@ namespace TrialByFire.Tresearch.Managers.Implementations
         private IMessageBank _messageBank { get; }
         private IAccountDeletionService _accountDeletionService { get; }
         private BuildSettingsOptions _options { get; }
+        private IAccountVerificationService _accountVerificationService { get; }
 
-        public AccountDeletionManager(ISqlDAO sqlDAO, ILogService logService, IMessageBank messageBank, IAccountDeletionService accountDeletionService, IOptionsSnapshot<BuildSettingsOptions> options)
+        public AccountDeletionManager(ISqlDAO sqlDAO, ILogService logService, IMessageBank messageBank, IAccountDeletionService accountDeletionService, IOptionsSnapshot<BuildSettingsOptions> options, IAccountVerificationService accountVerificationService)
         {
             _sqlDAO = sqlDAO;
             _logService = logService;
             _messageBank = messageBank;
             _accountDeletionService = accountDeletionService;
             _options = options.Value;
-
+            _accountVerificationService = accountVerificationService;
         }
 
 
@@ -40,36 +42,55 @@ namespace TrialByFire.Tresearch.Managers.Implementations
             try
             {
 
+
                 cancellationToken.ThrowIfCancellationRequested();
 
-                //checking if account is even there in the first place, if they are a guest then their username would be null.
+
                 if (Thread.CurrentPrincipal.Identity.Name.Equals(null))
                 {
                     return await _messageBank.GetMessage(IMessageBank.Responses.accountNotFound).ConfigureAwait(false);
                 }
 
+                string userName = Thread.CurrentPrincipal.Identity.Name;
                 string userAuthLevel = Thread.CurrentPrincipal.IsInRole("admin") ? "admin" : "user";
                 string admins = "";
+                string confirmed = "";
+
+                //verifying account here
+                IAccount account = new Account(userName, userAuthLevel);
+                confirmed = await _accountVerificationService.VerifyAccountAsync(account, cancellationToken).ConfigureAwait(false);
 
 
-
-                if (userAuthLevel.Equals("admin"))
+                if (confirmed.Equals(await _messageBank.GetMessage(IMessageBank.Responses.verifySuccess).ConfigureAwait(false)))
                 {
-                    admins = await _accountDeletionService.GetAmountOfAdminsAsync(cancellationToken).ConfigureAwait(false);
+                    if (userAuthLevel.Equals("admin"))
+                    {
+                        admins = await _accountDeletionService.GetAmountOfAdminsAsync(cancellationToken).ConfigureAwait(false);
 
-                    if (admins.Equals(await _messageBank.GetMessage(IMessageBank.Responses.getAdminsSuccess).ConfigureAwait(false)))
+                        if (admins.Equals(await _messageBank.GetMessage(IMessageBank.Responses.getAdminsSuccess).ConfigureAwait(false)))
+                        {
+                            return await _accountDeletionService.DeleteAccountAsync(cancellationToken).ConfigureAwait(false);
+                        }
+
+
+                        return await _messageBank.GetMessage(IMessageBank.Responses.getAdminsSuccess).ConfigureAwait(false);
+
+                    }
+                    else
                     {
                         return await _accountDeletionService.DeleteAccountAsync(cancellationToken).ConfigureAwait(false);
                     }
-
-
-                    return await _messageBank.GetMessage(IMessageBank.Responses.getAdminsSuccess).ConfigureAwait(false);
-
                 }
                 else
                 {
-                    return await _accountDeletionService.DeleteAccountAsync(cancellationToken).ConfigureAwait(false);
+                    //not sure if this is the right message to be sending back
+                    //return await _messageBank.GetMessage(IMessageBank.Responses.notConfirmed);
+                    return await _messageBank.GetMessage(IMessageBank.Responses.accountNotFound);
                 }
+
+
+
+
             }
 
             catch (OperationCanceledException)
